Create Image_Process_One_Directory.py
This process the images in a directory containing folder with images. This is designed to process all nested folder inside the main directory of the project. 

An example is the a Town/Municipality folder containing Barangay folders and containing drone images with different flights. 

When you run this script in agisoft metashape, make sure to put the two argument, the target directory where the folder of each flight is available and the directory for the ouput of orthomosaic.

Addition to this functionality, to eliminate the problem in storage system, this script will automatically delete the raw images after each process.